refactor(control-ui): split monolithic ControlUI into hooks and components
index.tsx had grown into a ~1400-line god component holding state
derivation, ~22 command handlers, hotkey registration, grid gestures,
sidebar filtering, styled-components and the whole layout composition.

Extract it into focused modules and leave index.tsx as a composition root
that only wires data and callbacks:

  hooks/useStreamFilters.ts   - filterStreams, filter text state, favorites
                                set, and the wall/live/other buckets
  hooks/useGridCommands.ts    - every send/stateDoc command wrapper
  hooks/useControlHotkeys.ts  - the audio-listen/blur/censor/swap/undo keys
  components/                 - ControlHeader, ControlGrid, ControlSidebar,
                                ControlStatusBar, ControlBanners,
                                StreamDelayBox
  layout.tsx                  - AppShell/Stack shell primitives

index.tsx drops from 1386 to 274 lines. View-state derivation already
lives in streamwallState.tsx, so it is reused rather than moved again, and
the package's public exports are unchanged.

Behavior is preserved verbatim, including the stable view-id command
addressing from #397, the destructive grid-shrink and preset-load confirm
guards, role gating, and the hotkey enableOnFormTags options.

Adds unit tests for the two new logic modules: useGridCommands (dispatch,
confirm guards, invite parse flow, Yjs write) and useStreamFilters (bucket
partitioning, kind/text filtering, favorites) - filterStreams was
previously untested.

Closes #393
